% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sambada.r
\name{changePath}
\alias{changePath}
\title{Adds folder to the PATH environmental variable}
\usage{
changePath(directory)
}
\arguments{
\item{directory}{character The path to samBada binaries folder}
}
\description{
Adds the \code{directory} path to the environmental PATH variable. This operation is only valid for the current R session. You must run change_path for every new R session. Alternatively, you can permantently edit your "PATH" environmental variable on your OS so that it entails the path to the binaries folder of sambada.
}
\author{
Solange Gaillard
}
