% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sambadaOnR.R
\docType{sambadaOnR}
\name{doc}
\alias{doc}
\title{sambadaOnR: A package for running sambada within R with pipeline from pre to post-processing}
\description{
The sambadaOnR package provides four categories of important functions:
Install sambada, Preprocessing, Running sambada and Post-processing.
}
\section{Install sambada functions}{

You can download sambada from GitHub using the function \code{\link{downloadSambada}}
}

\section{Preprocessing functions}{

The Preprocessing functions contain three functions: \code{prepareGeno}, \code{createEnv}, \code{prepareEnv}
}

\section{Running sambada function}{

To run sambada, you will want to use the function: \code{sambadaParallel}
}

\section{Postprocessing functions}{

The Postprocessing functions contain three functions: \code{prepareGeno}, \code{createEnv}, \code{prepareEnv}
}

