% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sambada.r
\name{downloadSambada}
\alias{downloadSambada}
\title{Download samBada}
\usage{
downloadSambada(directory = NULL)
}
\arguments{
\item{directory}{character The directory where sambada should be downloaded. If null, downloads in a (new) folder named sambada in the active directory.}
}
\description{
Downloads from GitHub the version of samBada that corresponds to your OS. Unzips the folder and adds the path to the binary folder to the environmental path variable. This operation is only valid for the current R session. You must run change_path for every new R session. Alternatively, you can manually edit your "PATH" environmental variable permanently on your OS so that it entails the path to the binaries folder of sambada (this procedure different for every OS).
}
\examples{
\dontrun{
downloadSamada('D:/Sambada')
}
}
\author{
Solange Duruz
}
