% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocessing.R
\name{plotResultInteractive}
\alias{plotResultInteractive}
\title{Interactive plotting of results}
\usage{
plotResultInteractive(preparedOutput, varEnv, envFile, species = NULL,
  pass = NULL, x = NULL, y = NULL, valueName = "pvalueG",
  chromo = "all", gdsFile = NULL, IDCol = NULL, popStrCol = NULL)
}
\arguments{
\item{preparedOutput}{char The prepared output list from prepare_output function}

\item{varEnv}{char The name of the environmental variable one wish to study (as in the header of \code{envFIle})}

\item{envFile}{char The file containing the input environmental variable of sambada.}

\item{species}{char The abbreviated latin name of the species without capitals nor punctuation (e.g. btaurus, chircus,...). Can be set to null if species not present in ensembl database}

\item{pass}{integer Number of BP around a SNP in which to look for an annotation in Ensembl. Set to null if species is null}

\item{x}{char The name of the column corresponding to the x-coordinate in the envFile. Can be set to null if unknown, in this case the maps will not be available}

\item{y}{char The name of the column corresponding to the y-coordinate in the env file. Can be set to null if x is null.}

\item{valueName}{char Name of the p- or q-value one wish to plot the manhattan on. This can be either pvalueG, pvalueW, qvalueG, qvalueW for G- or Waldscore respectively.}

\item{chromo}{char/integer Name or vector of name of the chromosome to investigate. If all is chosen (default), all numerical chromosome will be mapped. If your sambada output is large (typically if you are working with more than 50K genomic file), you should probably map a subset of your dataset (e.g. chr=1)}

\item{gdsFile}{char The GDS file created in the preprocessing of sambada. If null, will try with envFile(without -env.csv) and .gds}

\item{IDCol}{char The name of the column in envFile corresponding to the ID of the individuall. If provided, hover on the output map will give the id of the animal}

\item{popStrCol}{char The name or vector of name of column(s) in envFile describing population structure. If provided, additional layers on the map will be available reprenting population structure.}
}
\value{
None
}
\description{
Plots the manhattan plot for a given environmental variable. The plot is interactive and a map of the distribution of the marker can be retrieved as well as nearby genes listed in Ensembl.
}
\examples{
\dontrun{
plotResultInteractive('myFile','chircus',1,'Longitude','Latitude','bio1',c('1','2'))
}
}
\author{
Solange Gaillard
}
