% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{prepareGeno}
\alias{prepareGeno}
\title{Prepare genomic input}
\usage{
prepareGeno(filename, maf.thresh = NULL, missingness.thresh = NULL,
  LD.thresh = NULL, mgf.thresh = NULL, directory = NULL,
  interactiveChecks = FALSE, verbose = FALSE)
}
\arguments{
\item{filename}{char Name of the input file (must be in active directory). Can be .gds, .ped, .bed, .vcf. If different from .gds, a gds file (SNPrelate specific format) will be created unless no filtering options are chosen}

\item{maf.thresh}{double A number between 0 and 1 specifying the Major Allele Frequency (MAF) filtering (if null no filtering on MAF will be computed)}

\item{missingness.thresh}{double A number between 0 and 1 specifying the missing rate filtering (if null no filtering on missing rate will be computed)}

\item{LD.thresh}{double A number between 0 and 1 specifying the linkage desiquilibrium (LD) rate filtering (if null no filtering on LD will be computed)}

\item{mgf.thresh}{double A number between 0 and 1 specifying the Major Genotype Frequency (MGF) rate filtering (if null no filtering on MGF will be computed). NB: sambada computations relie on gentoypes}

\item{directory}{char The directory where binaries of sambada are saved. This parameter is not necessary if directoy path is permanently stored in the PATH environmental variable or if a function invoking samabada executable (prepareGeno or sambada_parallel) has been already run in the R active session.}

\item{interactiveChecks}{logical If TRUE, plots will show up showing distribution of allele frequency etc... and the user can interactively change the chosen threshold for maf.thresh, missingness.thresh, mgf.thresh (optional, default value=FALSE)}
}
\value{
None
}
\description{
Writes a new genomic file that sambada can work with after having applied the selected genomic filtering options. The output file has the same name as the input file but with a .csv extension
}
\examples{
prepareGeno('myPlinkFile.ped',maf.tresh=0.05, missingness.thresh=0.05,LD.thresh=0.2,mgf.thresh=0.8,interactiveChecks=TRUE)
prepareGeno('myGDSFile.gds',maf.tresh=0.05, missingness.thresh=0.05,LD.thresh=0.2,mgf.thresh=0.8,interactiveChecks=FALSE)
}
\author{
Solange Gaillard, Oliver Selmoni
}
