% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocessing.R
\name{prepareOutput}
\alias{prepareOutput}
\title{Prepare output (usefull for all postprocessing analysis)}
\usage{
prepareOutput(sambadaname, dimMax, gdsFile = NULL, popStr = FALSE,
  nrows = NULL, interactiveChecks = TRUE)
}
\arguments{
\item{sambadaname}{char The name of the genofile without extension name given to sambada (or outputfile of sambada without the ending -Out-Dim.csv)}

\item{dimMax}{integer The maximum number of dimension given in sambada}

\item{gdsFile}{char Name of the gds file associated with sambada's input file. If null, will try with \code{sambadaname}.gds}

\item{popStr}{logical Indicates whether sambada was run using the POPSTRVAR parameter (i.e. population structure was taken into account). Default false}

\item{nrows}{integer Specifies the number of line to read from the input file. Useful if saveType END ALL was used in sambada and that the number of models run is large so that the reading and processing is too slow. The saveType END parameter ensures that most significant models are located at the top of the file.}

\item{interactiveChecks}{logical}
}
\value{
a list containing a) $sambadaOutput a matrix containing the output from sambada with 3 additional column: corresponding snp, chromosome and position of the marker b) chrSNPNum The total number of SNPs in each chromosme c) $chrMaxPos The highest position found in each chromosome
}
\description{
Read sambada's output and prepare it by retrieving the snp position and chromosome (usefull for plotting manhattan)
}
\examples{
prepare_output('myFile',1)
}
