% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{prepareGeno}
\alias{prepareGeno}
\title{Prepare genomic input}
\usage{
prepareGeno(fileName, outputFile, saveGDS, mafThresh = NULL,
  missingnessThresh = NULL, ldThresh = NULL, mgfThresh = NULL,
  directory = NULL, interactiveChecks = FALSE, verbose = FALSE)
}
\arguments{
\item{fileName}{char Name of the input file (must be in active directory). Can be .gds, .ped, .bed, .vcf. If different from .gds, a gds file (SNPrelate specific format) will be created unless no filtering options are chosen}

\item{outputFile}{char Name of the output file. Must be a .csv}

\item{saveGDS}{logical If true (and if the input file extension is different from GDS) the GDS file will be saved. We recommend to set this parameter to TRUE to save time in subsequent functions that rely on GDS file}

\item{mafThresh}{double A number between 0 and 1 specifying the Major Allele Frequency (MAF) filtering (if null no filtering on MAF will be computed)}

\item{missingnessThresh}{double A number between 0 and 1 specifying the missing rate filtering (if null no filtering on missing rate will be computed)}

\item{ldThresh}{double A number between 0 and 1 specifying the linkage disequilibrium (LD) rate filtering (if null no filtering on LD will be computed)}

\item{mgfThresh}{double A number between 0 and 1 specifying the Major Genotype Frequency (MGF) rate filtering (if null no filtering on MGF will be computed). NB: sambada computations rely on genotypes}

\item{directory}{char The directory where binaries of sambada are saved. This parameter is not necessary if directory path is permanently stored in the PATH environmental variable or if a function invoking sambada executable (\code{prepareGeno} or \code{sambadaParallel}) has been already run in the R active session.}

\item{interactiveChecks}{logical If TRUE, plots will show up showing distribution of allele frequency etc... and the user can interactively change the chosen threshold for \code{mafThresh}, \code{missingnessThresh}, \code{mgfThresh} (optional, default value=FALSE)}

\item{verbose}{logical Turn on verbose mode}
}
\value{
None
}
\description{
Writes a new genomic file that sambada can work with after having applied the selected genomic filtering options. The output file has the same name as the input file but with a .csv extension
}
\examples{
\dontrun{
#With ped input file
prepareGeno('myPlinkFile.ped','mySambadaFile.csv',TRUE, mafThresh=0.05, 
     missingnessThresh=0.05, mgfThresh=0.8,interactiveChecks=TRUE)

#With gds input file
prepareGeno('myGDSFile.gds','mySambadaFile.csv',FALSE, mafThresh=0.05, 
     missingnessThresh=0.05,mgfThresh=0.8,interactiveChecks=FALSE)
}
}
\author{
Solange Duruz, Oliver Selmoni
}
